+++ next/src/lib/processUrl.ts
@@ -1,20 +1,21 @@
 export const BASE_URL = 'https://next-sanity-template-demo.vercel.app'
 
-type Options = {
-	base?: boolean
-	params?: string
-}
-
 export default function (
 	page: Sanity.PageBase,
-	{ base = true, params }: Options = {},
+	{
+		base = true,
+		params,
+	}: {
+		base?: boolean
+		params?: string
+	} = {},
 ) {
 	// prettier-ignore
 	const directory =
 		page._type === 'blog.post' ? 'blog' :
 		null
 
-	const slug = page.metadata?.slug.current
+	const slug = page.metadata?.slug?.current
 	const path = slug === 'index' ? null : slug
 
 	return (