+++ next/src/ui/modules/CreativeModule/CustomHTMLSubmodule.tsx
@@ -1,0 +1,14 @@
+import CustomHTML from '../CustomHTML'
+
+export type CustomHTMLSubmoduleType = Sanity.Module<'custom-html'> &
+	Partial<{
+		html: string
+	}>
+
+export default function CustomHTMLSubmodule({
+	module,
+}: {
+	module: CustomHTMLSubmoduleType
+}) {
+	return <CustomHTML html={module.html} />
+}