+++ sanity/schemas/modules/creative/index.ts
@@ -34,7 +34,13 @@ export default defineType({
 						defineField({
 							name: 'subModules',
 							type: 'array',
-							of: [creativeCtas, creativeIcon, creativeImage, creativeRichtext],
+							of: [
+								creativeCtas,
+								creativeIcon,
+								creativeImage,
+								creativeRichtext,
+								{ type: 'custom-html' },
+							],
 						}),
 						defineField({
 							name: 'colSpan',